perf(translator): i64 bitwise flavor + f64 multiplication
Bitwise ops (& | ^ << >> ~) yield a ToInt32 result sign-extended to i64,
so a binding fed a bitwise result keeps i64 instead of round-tripping
through f64 per op. The levenshtein Myers inner loop (~12 bitwise ops per
char) halves: 162 ms -> 78 ms, beating node (130) and bun (123). LLVM
does not fold the inner-loop f64<->i32 cast chain on its own; values stay
under 2^31 so i64 matches ES number exactly (checksum 600000).

Multiplication (*) joins / and ** as always-f64: a product's intermediate
can exceed 2^53 even for small operands (a 31-bit LCG seed * 1103515245
~= 2.4e18), where exact i64 diverges from the rounded ES number. This
restores the loop-data-dependent checksum to match node/bun. +/-/% keep
the integral combine (intermediates stay under 2^53 for the loop counters
flavor inference targets).

binary.rs arithmetic emit now treats * like / (operands at f64), matching
flavor.rs's binary_flavor; is_number_expr recognizes bitwise/~ results
non-recursively for cast coordination.

+++ crates/dashscript/src/translator/expressions/unary.rs
@@ -48,15 +48,17 @@ pub(super) fn unary_expr(un: &UnaryExpression, ctx: &Ctx<'_>) -> Expr {
                 })
             }
         }
-        // `~a` → `!ToInt32(a) as f64` (TS `~` is 32-bit bitwise NOT). The
-        // operand casts via `bitwise_operand` (f64 → i64 → i32 for the JS
-        // `ToInt32` wrap; i64 skips the hop); bound to a local so `as` never
-        // binds into a compound operand.
+        // `~a` → `!ToInt32(a)` sign-extended to `i64` (a `ToInt32` result is a
+        // signed i32, exact in i64), so a binding fed `~a` keeps `i64` flavor
+        // instead of round-tripping through `f64`. The operand casts via
+        // `bitwise_operand` (f64 → i64 → i32 for the JS `ToInt32` wrap; i64
+        // skips the hop); bound to a local so `as` never binds into a compound
+        // operand.
         UnaryOperator::BitwiseNot => {
             let a = super::bitwise_operand(&un.argument, ctx, true);
             parse_quote!({
                 let __a = #a;
-                (!__a) as f64
+                (!__a) as i64
             })
         }
         // `typeof x` is a compile-time type query (DashScript is statically
